third_party/WebKit/LayoutTests/fast/events/inputevents/inputevent-keyboard.html

<!DOCTYPE html>
<html>
<head>
<title>InputEvent: beforeinput inputType</title>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
</head>
<body>
<input type="text" id="txt">
<script>
test(function() {
    var lastBeforeInputType = '';
    var lastInputType = '';
    var txt = document.getElementById('txt');
    txt.addEventListener('beforeinput', function(event) {
        assert_true(event instanceof InputEvent);
        assert_false(event.isComposing);
        lastBeforeInputType = event.inputType;
    });
    txt.addEventListener('input', function(event) {
        assert_true(event instanceof InputEvent);
        assert_false(event.isComposing);
        lastInputType = event.inputType;
    });
    if (!window.eventSender) {
        document.write('This test requires eventSender');
    } else {
        const NO_INPUT_EVENT_FIRED = 'NO_INPUT_EVENT_FIRED';
        function testKeyDownInputType(key, modifiers, beforeInputType, inputType) {
            inputType = inputType || beforeInputType;
            lastBeforeInputType = NO_INPUT_EVENT_FIRED;
            lastInputType = NO_INPUT_EVENT_FIRED;
            eventSender.keyDown(key, modifiers);
            assert_equals(lastBeforeInputType, beforeInputType, `${modifiers.toString()}+${key} should produce beforeInputType: ${inputType}`);
            assert_equals(lastInputType, inputType, `${modifiers.toString()}+${key} should produce inputType: ${inputType}`);
        }

        txt.focus();
        // Typing
        testKeyDownInputType('a', [], 'insertText');
        testKeyDownInputType('6', [], 'insertText');
        testKeyDownInputType('Backspace', [], 'deleteContentBackward');
        testKeyDownInputType('l', ['shiftKey'], 'insertText');
        testKeyDownInputType('w', ['shiftKey'], 'insertText');
        // TODO(chongz): Add tests for Enter key on <textarea> and ContentEditable.
        testKeyDownInputType('Enter', [], 'insertLineBreak', NO_INPUT_EVENT_FIRED);
        testKeyDownInputType('Enter', ['shiftKey'], 'insertLineBreak', NO_INPUT_EVENT_FIRED);

        // Keyboard commands
        var isMacOS = (navigator.userAgent.indexOf('Mac OS X') != -1);
        if (!isMacOS) {
            // MacOS's eventSender does not work on hot keys other than arrows.
            testKeyDownInputType('z', ['ctrlKey'], 'undo');
            testKeyDownInputType('z', ['ctrlKey', 'shiftKey'], 'redo');
        }
        // Move command should not generate input events.
        testKeyDownInputType('ArrowLeft', [], NO_INPUT_EVENT_FIRED);
        testKeyDownInputType('ArrowLeft', ['shiftKey'], NO_INPUT_EVENT_FIRED);
        testKeyDownInputType('Home', [], NO_INPUT_EVENT_FIRED);
    }
}, 'Testing beforeinput inputType');
</script>
</body>
</html>
